backend/requirements.txt
fastapi
uvicorn
python-multipart
spacy
PyMuPDF
python-dotenv